Set time on time change
